Add ConfigManager to AppServiceProvider
This is a breaking change, but it's a good one.

The ConfigManager is now available as a singleton, so it can be used
anywhere in the application.

This is a breaking change because the ConfigManager is now a singleton,
so it can no longer be instantiated.

This is a good change because it makes the ConfigManager more useful.